refactor(template): exports, refactor scheduling folder
