+++ libs/template/src/index.ts
@@ -1,7 +1,6 @@
-export * from './lib/core';
-export * from './lib/render-strategies';
 export { PushPipe, PushModule } from './lib/push';
 export { LetDirective, LetModule } from './lib/let';
+// EXPERIMENTAL
 export {
   UnpatchEventsDirective,
   UnpatchEventsModule
