+++ libs/template/src/lib/render-strategies/rxjs/scheduling/animationFrameTick.ts
@@ -2,7 +2,7 @@ import { Observable, SchedulerLike, Subscription } from 'rxjs';
 import { getZoneUnPatchedApi } from '../../../core';
 
 export const animationFrameTick = () =>
-  new Observable(subscriber => {
+  new Observable<number>(subscriber => {
     const id = getZoneUnPatchedApi('requestAnimationFrame')(() => {
       subscriber.next(0);
       subscriber.complete();
