+++ libs/template/src/lib/render-strategies/rxjs/scheduling/idleCallbackTick.ts
@@ -2,7 +2,7 @@ import { Observable } from 'rxjs';
 import { cancelIdleCallback, requestIdleCallback } from './idleCallback';
 
 export const idleCallbackTick = () =>
-  new Observable(subscriber => {
+  new Observable<number>(subscriber => {
     const id = requestIdleCallback(() => {
       subscriber.next(0);
       subscriber.complete();
