+++ libs/template/src/lib/render-strategies/rxjs/scheduling/index.ts
@@ -1,6 +1,6 @@
-export { unpatchedAsapScheduler } from './asapScheduler';
-export { getPostTaskScheduler } from './getPostTaskScheduler';
-export { idleScheduler } from './idleCallbackScheduler';
-export { unpatchedAnimationFrameScheduler } from './animationFrameScheduler';
+export * from './promiseTick';
+export * from './postTaskTick';
+export * from './idleCallbackTick';
+export * from './animationFrameTick';
 export * from './interfaces';
-export * from './priority-scheduler-map';
+export * from './priority-tick-map';
