+++ libs/template/src/lib/render-strategies/strategies/index.ts
@@ -1,1 +1,5 @@
 export { getStrategies } from './strategies-map';
+export { getLocalStrategies } from './local.strategy';
+export { getGlobalStrategies } from './global.strategy';
+export { createNativeStrategy } from './native.strategy';
+export { createNoopStrategy } from './noop.strategy';
